Explain briefly the need for pre-merge
When I first read this section, I was scared by the "This feature requires a complete understanding of normalization and full stack operation" and thought that it was and advanced feature that I did not need try understanding anytime soon. The consequence it that it took me longer to discover that this was the correct solution for not breaking a child router when loading data. I think it is valuable to briefly mention the purpose already here so the reader will get an idea when s/he needs to come back to learn more.