Partial Routes: can also use :will-mount
For my case, specifying the path to the leaf target in the closet ancestor with a deferred will-enter is a better match then either globally at start or forcing its siblings to know about its internal structure. But it wasn't obvious that this is possibly so I think it is worth documenting it.