Update DevelopersGuide.adoc
Add cider/cider-nrepl dependency to starter and bump to latest dependency versions

Go To Definition functionality in VSCode using Calva does not work if cider/cider-nrepl is not included in the dependencies.

If you start shadow-cljs from VSCode it will explicitly include it by running
`npx shadow-cljs -d cider/cider-nrepl:0.26.0 watch :main`

Since this guide suggests starting shadow-cljs in server mode this dependency needs to be included in deps or perhaps in the run command later eg `npx shadow-cljs -d cider/cider-nrepl:0.26.0 server`

I also noticed that the Tip below to configure clj-kondo required quoting the lint argument or zsh, but it also did not appear to do anything when I ran it... 

`
➜  first-project git:(master) ✗ cat .clj-kondo/config.edn
{}
➜  first-project git:(master) ✗ clj-kondo --copy-configs --dependencies --lint '(clojure -Spath)'
➜  first-project git:(master) ✗ cat .clj-kondo/config.edn
{}
➜  first-project git:(master) ✗ clj-kondo --copy-configs --dependencies --lint  (clojure -Spath)
zsh: number expected
➜  first-project git:(master) ✗ clj-kondo --copy-configs --dependencies --lint '(clojure -Spath)'
[clj-kondo] Auto-loading config path: com.fulcrologic/fulcro
➜  first-project git:(master) ✗ cat .clj-kondo/config.edn
{}
`
VSCode however is not complaining about defsc or anything so it either this step is no longer necessary (I think clj-kondo can discover config in libraries and now just works) or this command updated something other than the config file.  
Pretty sure this is no longer needed so I removed it... if still needed then perhaps it can be updated to quote the lint argument and better document what the expected result of this is.